pyproject.toml
[project]
name = "model-traffic-manager"
version = "0.1.0"
description = "Policy-driven AI traffic router for Azure and AKS"
requires-python = "==3.12.*"
dependencies = [
    "anyio==4.12.1",
    "azure-core==1.38.3",
    "azure-identity==1.25.3",
    "fastapi==0.135.1",
    "httpx==0.28.1",
    "opentelemetry-api==1.40.0",
    "opentelemetry-instrumentation-fastapi==0.61b0",
    "opentelemetry-instrumentation-httpx==0.61b0",
    "opentelemetry-sdk==1.40.0",
    "prometheus-client==0.24.1",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pyyaml==6.0.3",
    "redis==7.3.0",
    "structlog==25.5.0",
    "uvicorn==0.41.0",
]

[dependency-groups]
dev = [
    "mypy==1.19.1",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "respx==0.22.0",
    "ruff==0.15.6",
    "types-pyyaml==6.0.12.20250915",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
packages = ["app"]
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
